Add close messages when commands are interrupted
When a command is interrupted, it will notify its owner of the
interruption with a message identifying the command. This will prevent
calls waiting until timeout (potentially forever) while waiting for a
response.

The on_close role callbacks are now properly called on termination.

Note: the Message.receive_macro now includes a clause for handling these
messages, but it might be improved to take advantage of more specific
matching.

This commit also includes some cleanup of the Publisher role -- it was
carrying around unnecessary state.